move getAvatar into lens-protocol
