+++ apps/app/src/components/Causes/CausePage.tsx
@@ -9,8 +9,7 @@ import { useTranslation } from 'react-i18next'
 
 import { CURRENCIES } from '@/constants'
 import { formatLocation } from '@/lib/formatLocation'
-import getAvatar from '@/lib/getAvatar'
-import { lensClient, usePublication } from '@/lib/lens-protocol'
+import { getAvatar, lensClient, usePublication } from '@/lib/lens-protocol'
 import {
   CauseMetadata,
   CauseMetadataBuilder,
