+++ apps/app/src/components/Dashboard/VolunteerDashboard/VolunteerHome.tsx
@@ -10,8 +10,7 @@ import { useTranslation } from 'react-i18next'
 import { GridItemTwelve, GridLayout } from '@/components/GridLayout'
 import Slug from '@/components/Shared/Slug'
 import { Card } from '@/components/UI/Card'
-import getAvatar from '@/lib/getAvatar'
-import { lensClient } from '@/lib/lens-protocol'
+import { getAvatar, lensClient } from '@/lib/lens-protocol'
 import { isPost } from '@/lib/metadata'
 import { useAppPersistStore } from '@/store/app'
