+++ apps/app/src/components/VHRs/index.tsx
@@ -4,9 +4,8 @@ import { NextPage } from 'next'
 import { SetStateAction, useEffect, useState } from 'react'
 
 import { CORS_PROXY, VHR_TOP_HOLDERS_URL } from '@/constants'
-import getAvatar from '@/lib/getAvatar'
 import isVerified from '@/lib/isVerified'
-import { getProfilesOwnedBy } from '@/lib/lens-protocol'
+import { getAvatar, getProfilesOwnedBy } from '@/lib/lens-protocol'
 
 import { GridItemTwelve, GridLayout } from '../GridLayout'
 import SEO from '../utils/SEO'
